refactor: extract useCommandCodeAuth hook (#3501 Phase 1h)
