+++ src/app/(dashboard)/dashboard/providers/[id]/ProviderDetailPageClient.tsx
@@ -5,6 +5,8 @@ import { useState, useEffect, useCallback, useRef, useMemo } from "react";
 import { useProviderConnections } from "./hooks/useProviderConnections";
 import { useProviderSettings } from "./hooks/useProviderSettings";
 import { useProviderModels } from "./hooks/useProviderModels";
+// Phase 1h: commandCode auth flow extracted to hooks/useCommandCodeAuth.ts
+import { useCommandCodeAuth } from "./hooks/useCommandCodeAuth";
 // Phase 1g: ProviderPlaygroundPanel + helpers extracted to components/ProviderPlaygroundPanel.tsx
 import ProviderPlaygroundPanel from "./components/ProviderPlaygroundPanel";
 import { useNotificationStore } from "@/store/notificationStore";
@@ -103,7 +105,7 @@ import {
   formatProviderModelsErrorResponse,
   type ProviderMessageTranslator,
   type LocalProviderMetadata,
-  type CommandCodeAuthFlowState,
+  // CommandCodeAuthFlowState moved to hooks/useCommandCodeAuth.ts (Phase 1h)
   type CompatByProtocolMap,
   type CompatModelRow,
   type CompatModelMap,
@@ -156,14 +158,6 @@ export default function ProviderDetailPageClient() {
   const [showSiliconFlowEndpointModal, setShowSiliconFlowEndpointModal] = useState(false);
   const [siliconFlowInitialBaseUrl, setSiliconFlowInitialBaseUrl] = useState<string | undefined>();
   const [showRiskNoticeModal, setShowRiskNoticeModal] = useState(false);
-  const [commandCodeAuthState, setCommandCodeAuthState] = useState<CommandCodeAuthFlowState>({
-    phase: "idle",
-    state: "",
-    authUrl: "",
-    callbackUrl: "",
-    expiresAt: null,
-    message: "",
-  });
   const [showEditModal, setShowEditModal] = useState(false);
   const [showEditNodeModal, setShowEditNodeModal] = useState(false);
   const [showTutorialModal, setShowTutorialModal] = useState(false);
@@ -223,8 +217,6 @@ export default function ProviderDetailPageClient() {
   );
   const [exportingGeminiAuthId, setExportingGeminiAuthId] = useState<string | null>(null);
   const [importGeminiModalOpen, setImportGeminiModalOpen] = useState(false);
-  const commandCodeAuthWindowRef = useRef<Window | null>(null);
-  const commandCodeAuthTimerRef = useRef<number | null>(null);
   const pendingRiskActionRef = useRef<(() => void) | null>(null);
   const { acknowledged: riskAcknowledged, acknowledge: acknowledgeRisk } =
     useRiskAcknowledged(providerId);
@@ -748,257 +740,21 @@ export default function ProviderDetailPageClient() {
     setShowRiskNoticeModal(false);
   }, []);
 
-  const clearCommandCodeAuthTimer = useCallback(() => {
-    if (commandCodeAuthTimerRef.current !== null) {
-      window.clearTimeout(commandCodeAuthTimerRef.current);
-      commandCodeAuthTimerRef.current = null;
-    }
-  }, []);
-
-  useEffect(() => {
-    return () => {
-      clearCommandCodeAuthTimer();
-      commandCodeAuthWindowRef.current?.close?.();
-    };
-  }, [clearCommandCodeAuthTimer]);
-
-  const handleCloseAddApiKeyModal = useCallback(() => {
-    clearCommandCodeAuthTimer();
-    setSiliconFlowInitialBaseUrl(undefined);
-    commandCodeAuthWindowRef.current?.close?.();
-    commandCodeAuthWindowRef.current = null;
-    setCommandCodeAuthState({
-      phase: "idle",
-      state: "",
-      authUrl: "",
-      callbackUrl: "",
-      expiresAt: null,
-      message: "",
-    });
-    setShowAddApiKeyModal(false);
-  }, [clearCommandCodeAuthTimer]);
-
-  const handleCommandCodeAuthApply = useCallback(
-    async (state: string, connectionId?: string, name?: string, setDefault?: boolean) => {
-      setCommandCodeAuthState((current) => ({
-        ...current,
-        phase: "applying",
-        message: "Applying browser-approved key…",
-      }));
-
-      try {
-        const res = await fetch("/api/providers/command-code/auth/apply", {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ state, connectionId, name, setDefault }),
-        });
-        const data = await res.json().catch(() => ({}));
-
-        if (!res.ok) {
-          const errorMessage = data.error || "Failed to apply Command Code auth";
-          setCommandCodeAuthState((current) => ({
-            ...current,
-            phase: "error",
-            message: errorMessage,
-          }));
-          notify.error(errorMessage);
-          return false;
-        }
-
-        setCommandCodeAuthState((current) => ({
-          ...current,
-          phase: "applied",
-          message: "Command Code connected",
-        }));
-        commandCodeAuthWindowRef.current?.close?.();
-        commandCodeAuthWindowRef.current = null;
-        await fetchConnections();
-        handleCloseAddApiKeyModal();
-        notify.success("Command Code connection added");
-        return true;
-      } catch (error) {
-        console.error("Error applying Command Code auth:", error);
-        setCommandCodeAuthState((current) => ({
-          ...current,
-          phase: "error",
-          message: "Failed to apply Command Code auth",
-        }));
-        notify.error("Failed to apply Command Code auth");
-        return false;
-      }
-    },
-    [fetchConnections, handleCloseAddApiKeyModal, notify]
-  );
-
-  const handleStartCommandCodeAuth = useCallback(async () => {
-    if (commandCodeAuthState.phase === "starting" || commandCodeAuthState.phase === "polling") {
-      return;
-    }
-
-    clearCommandCodeAuthTimer();
-    commandCodeAuthWindowRef.current?.close?.();
-
-    const popup = window.open("about:blank", "_blank");
-    setCommandCodeAuthState({
-      phase: "starting",
-      state: "",
-      authUrl: "",
-      callbackUrl: "",
-      expiresAt: null,
-      message: "Opening Command Code Studio…",
-    });
-
-    try {
-      const res = await fetch("/api/providers/command-code/auth/start", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-      });
-      const data = await res.json().catch(() => ({}));
-
-      if (!res.ok || !data.state || !data.authUrl) {
-        const errorMessage = data.error || "Failed to start Command Code auth";
-        setCommandCodeAuthState((current) => ({
-          ...current,
-          phase: "error",
-          message: errorMessage,
-        }));
-        notify.error(errorMessage);
-        popup?.close?.();
-        return;
-      }
-
-      setCommandCodeAuthState({
-        phase: "polling",
-        state: data.state,
-        authUrl: data.authUrl,
-        callbackUrl: data.callbackUrl || "",
-        expiresAt: data.expiresAt || null,
-        message: "Open the auth URL, approve access, then paste the returned key/JSON/URL below…",
-      });
-
-      if (popup) {
-        try {
-          popup.opener = null;
-        } catch {
-          // Ignore opener cleanup failures.
-        }
-        popup.location.href = data.authUrl;
-        commandCodeAuthWindowRef.current = popup;
-      } else {
-        const fallbackPopup = window.open(data.authUrl, "_blank", "noopener,noreferrer");
-        if (!fallbackPopup) {
-          setCommandCodeAuthState((current) => ({
-            ...current,
-            phase: "error",
-            message: "Popup blocked. Please allow popups and try Command Code Connect again.",
-          }));
-          notify.error("Popup blocked. Please allow popups and try Command Code Connect again.");
-          return;
-        }
-        commandCodeAuthWindowRef.current = fallbackPopup;
-      }
-
-      const deadline = data.expiresAt ? new Date(data.expiresAt).getTime() : Date.now() + 180000;
-      const poll = async () => {
-        if (Date.now() >= deadline) {
-          setCommandCodeAuthState((current) => ({
-            ...current,
-            phase: "expired",
-            message: "Command Code link expired",
-          }));
-          commandCodeAuthWindowRef.current?.close?.();
-          commandCodeAuthWindowRef.current = null;
-          notify.error("Command Code auth expired");
-          clearCommandCodeAuthTimer();
-          return;
-        }
-
-        try {
-          const statusRes = await fetch(
-            `/api/providers/command-code/auth/status?state=${encodeURIComponent(data.state)}`,
-            { method: "GET", cache: "no-store" }
-          );
-          const statusData = await statusRes.json().catch(() => ({}));
-          const status = String(statusData.status || statusData.state || statusData.phase || "")
-            .toLowerCase()
-            .trim();
-
-          if (status === "expired") {
-            setCommandCodeAuthState((current) => ({
-              ...current,
-              phase: "expired",
-              message: "Command Code link expired",
-            }));
-            commandCodeAuthWindowRef.current?.close?.();
-            commandCodeAuthWindowRef.current = null;
-            notify.error("Command Code auth expired");
-            clearCommandCodeAuthTimer();
-            return;
-          }
-
-          if (status === "applied") {
-            setCommandCodeAuthState((current) => ({
-              ...current,
-              phase: "applied",
-              message: "Command Code connected",
-            }));
-            commandCodeAuthWindowRef.current?.close?.();
-            commandCodeAuthWindowRef.current = null;
-            await fetchConnections();
-            handleCloseAddApiKeyModal();
-            notify.success("Command Code connection added");
-            clearCommandCodeAuthTimer();
-            return;
-          }
-
-          if (status === "received") {
-            setCommandCodeAuthState((current) => ({
-              ...current,
-              phase: "received",
-              message: "Browser approved, applying…",
-            }));
-            clearCommandCodeAuthTimer();
-            await handleCommandCodeAuthApply(
-              data.state,
-              statusData.connectionId,
-              statusData.name,
-              statusData.setDefault
-            );
-            return;
-          }
-        } catch {
-          // Keep polling until the contract reports a terminal state or timeout.
-        }
-
-        commandCodeAuthTimerRef.current = window.setTimeout(poll, 2000);
-      };
-
-      commandCodeAuthTimerRef.current = window.setTimeout(poll, 1000);
-    } catch (error) {
-      console.error("Error starting Command Code auth:", error);
-      setCommandCodeAuthState((current) => ({
-        ...current,
-        phase: "error",
-        message: "Failed to start Command Code auth",
-      }));
-      notify.error("Failed to start Command Code auth");
-      popup?.close?.();
-      commandCodeAuthWindowRef.current = null;
-      clearCommandCodeAuthTimer();
-    }
-  }, [
+  // ── Phase 1h: commandCode auth flow ─────────────────────────────────────
+  const {
+    commandCodeAuthState,
     clearCommandCodeAuthTimer,
     handleCloseAddApiKeyModal,
-    commandCodeAuthState.phase,
-    fetchConnections,
     handleCommandCodeAuthApply,
+    handleStartCommandCodeAuth,
+    handleOpenCommandCodeConnect,
+  } = useCommandCodeAuth({
+    providerId,
+    fetchConnections,
+    setSiliconFlowInitialBaseUrl,
+    setShowAddApiKeyModal,
     notify,
-  ]);
-
-  const handleOpenCommandCodeConnect = useCallback(() => {
-    setShowAddApiKeyModal(true);
-    void handleStartCommandCodeAuth();
-  }, [handleStartCommandCodeAuth]);
+  });
 
   const handleSaveApiKey = async (formData) => {
     try {
